@@ -18,13 +18,18 @@
     "@angular/common": "^6.0.0",
     "@angular/compiler": "^6.0.0",
     "@angular/core": "^6.0.0",
+    "@angular/flex-layout": "^6.0.0-beta.16",
     "@angular/forms": "^6.0.0",
     "@angular/http": "^6.0.0",
     "@angular/material": "6.0.0",
     "@angular/platform-browser": "^6.0.0",
     "@angular/platform-browser-dynamic": "^6.0.0",
     "@angular/router": "^6.0.0",
+    "@auth0/angular-jwt": "^2.0.0",
+    "angular-bootstrap-md": "^6.2.6",
+    "bcryptjs": "^2.4.3",
     "body-parser": "^1.18.2",
+    "chart.js": "^2.5.0",
     "cookie-parser": "^1.4.3",
     "core-js": "^2.5.5",
     "csurf": "^1.9.0",
@@ -32,20 +37,22 @@
     "express": "^4.16.3",
     "express-handlebars": "^3.0.0",
     "express-session": "^1.15.6",
+    "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "handlebars": "^4.0.11",
     "handlebars-helpers": "^0.10.0",
     "handlebars-layouts": "^3.1.4",
     "hbs": "~4.0.1",
+    "jsonwebtoken": "^8.3.0",
     "mongoose": "^5.0.17",
     "morgan": "~1.9.0",
-    "rxjs": "^6.3.3",
+    "rxjs": "6.2.2",
     "serve-favicon": "~2.5.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.7.0",
-    "@angular/cli": "~6.1.1",
+    "@angular-devkit/build-angular": "^0.8.3",
+    "@angular/cli": "^6.2.3",
     "@angular/compiler-cli": "^6.1.0",
     "@angular/language-service": "^6.1.0",
     "@types/jasmine": "~2.8.3",
@@ -59,9 +66,9 @@
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
+    "protractor": "^5.4.1",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "2.7.2"
+    "typescript": "2.9.1"
   }
 }
