@@ -1,5 +1,12 @@
+// if (!process.env.NODE_ENV) process.env.NODE_ENV = 'development';
+
+// const env = process.env.NODE_ENV;
+
 var jwt = require('jsonwebtoken');
-var config = require('../config');
+// var config = require('../config');
+// var config = require(`../../../config/config.${env}.js`);
+var config = require(`./configLoader`).secret;
+
 
 function verifyToken(req, res, next) {
     var token = req.headers['x-access-token'];
