@@ -2,6 +2,7 @@
 const mongoose = require('mongoose'),
     Customer = require('../models/customer'),
     State = require('../models/state'),
+    City = require('../models/city'),
     Project = require('../models/project'),
     dbConfig = require('./configLoader').databaseConfig,
     connectionString = `mongodb://${dbConfig.host}/${dbConfig.database}`,
@@ -192,8 +193,8 @@ class DBSeeder {
 
         //States
         var states = [
-            { "name": "Alabama", "abbreviation": "AL" },
-            { "name": "Montana", "abbreviation": "MT" },
+            { "name": "Gujarat", "abbreviation": "GJ" },
+            { "name": "Maharastra", "abbreviation": "MH" },
             { "name": "Alaska", "abbreviation": "AK" },
             { "name": "Nebraska", "abbreviation": "NE" },
             { "name": "Arizona", "abbreviation": "AZ" },
@@ -252,6 +253,24 @@ class DBSeeder {
         for (i = 0; i < l; i++) {
             var state = new State({ 'id': i + 1, 'name': states[i].name, 'abbreviation': states[i].abbreviation });
             state.save();
+        }
+
+        //Cities
+
+        var cities =
+            [
+                { "name": "Baroda", "abbreviation": "BRC" },
+                { "name": "Mumbai", "abbreviation": "BCT" }
+            ];
+
+        var l = cities.length,
+            i;
+
+        City.remove({});
+        var state = { 'id': 1, 'abbreviation': 'GJ', 'name': 'Gujarat' };
+        for (i = 0; i < l; i++) {
+            var city = new City({ 'id': i + 1, 'name': cities[i].name, 'abbreviation': cities[i].abbreviation, 'state': state, 'stateId': 1 });
+            city.save();
         }
 
         //Projects
