@@ -8,13 +8,27 @@ import { CoreModule } from './core/core.module';
 import { SharedModule } from './shared/shared.module';
 import { DemoComponent } from './demo/demo.component';
 import { MaterialModule } from './shared/material/material.module';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { MDBBootstrapModule } from 'angular-bootstrap-md';
+import { JwtModule } from '@auth0/angular-jwt';
+import { HttpClientModule } from '@angular/common/http';
+
 
 @NgModule({
   imports: [
     BrowserModule,
-    BrowserAnimationsModule,  
+    BrowserAnimationsModule,
+    HttpClientModule,
+    JwtModule.forRoot({
+      config: {
+        tokenGetter: () => {
+          return localStorage.getItem('access_token');
+        },
+        whitelistedDomains: ['localhost:3001'],
+        blacklistedRoutes: ['localhost:3001/auth/']
+      }
+    }),
     MaterialModule,
+    MDBBootstrapModule,
     AppRoutingModule,
     CoreModule,     //Singleton objects
     SharedModule  //Shared (multi-instance) objects
@@ -24,6 +38,7 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
     AppRoutingModule.components,
     DemoComponent
   ],
+
   bootstrap: [AppComponent]
 })
 export class AppModule { }
